Change the way read_file and write_file works
Until now, the activity do not modified the real activities sources,
but a copy in the instance directory. At saving, Develop created a bundle
with the new version of the modified files. This metodology was safe,
but doing tests was complex and the code was complex too.
Now Develop activity will save the code in the Activities directory.
Safety will be solved using git or cloning the activity with the
new Sugar functionality.
write_file save only a json file with session data, the bundle will be
created by the user with a new button when needed. This functionality
is not available right now.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>